move comon code to util test class
