+++ src/test/java/tech/stackable/gis/hbase/filter/TestWithinFilter.java
@@ -1,10 +1,7 @@
 package tech.stackable.gis.hbase.filter;
 
-import ch.hsr.geohash.GeoHash;
-import com.google.common.base.Splitter;
 import com.vividsolutions.jts.geom.Geometry;
 import com.vividsolutions.jts.io.WKTReader;
-import org.apache.commons.collections.iterators.ArrayIterator;
 import org.apache.hadoop.hbase.*;
 import org.apache.hadoop.hbase.client.Durability;
 import org.apache.hadoop.hbase.client.Put;
@@ -13,32 +10,20 @@ import org.apache.hadoop.hbase.filter.FilterList;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import tech.stackable.gis.hbase.AbstractTestUtil;
 
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
 import java.math.BigInteger;
-import java.util.*;
 
 import static org.junit.Assert.assertEquals;
 
-public class TestWithinFilter extends AbstractTestFilter {
+public class TestWithinFilter extends AbstractTestUtil {
     public static final int POINT_COUNT = 10;
     public static final String TABLE = "TestWithinFilter";
-    private static final byte[] FAMILY_A = "a".getBytes();
     private static final byte[] FAMILY_B = "b".getBytes();
-    private static final String[] COLUMNS = new String[]{
-            "lon", "lat", "id", "name", "address",
-            "city", "url", "phone", "type", "zip"};
-    private static final ArrayIterator COLS = new ArrayIterator(COLUMNS);
-    private static final Splitter SPLITTER = Splitter.on('\t')
-            .trimResults()
-            .limit(COLUMNS.length);
     private static final byte[] X_COL = "lon".getBytes();
     private static final byte[] Y_COL = "lat".getBytes();
     private static final byte[][] COLUMNS_SCAN = {"id".getBytes(), X_COL, Y_COL};
     private static final byte[][] COLUMNS_RECTANGLE_CHECK = {X_COL, Y_COL};
-    public static int WIFI_COUNT;
 
     @BeforeClass
     public static void before() throws Exception {
@@ -54,55 +39,6 @@ public class TestWithinFilter extends AbstractTestFilter {
 
         load_wifi_data();
         load_rectangle_check_data();
-    }
-
-    private static void load_wifi_data() throws IOException {
-        BufferedReader reader = new BufferedReader(new FileReader("src/test/resources/wifi_4326.txt"));
-        reader.readLine(); // ignore header
-        String line;
-
-        int records = 0, duplicates = 0;
-        Set<String> uniqueKeys = new HashSet<>();
-        long start = System.currentTimeMillis();
-
-        try {
-            while ((line = reader.readLine()) != null) {
-                COLS.reset();
-                Iterator<String> vals = SPLITTER.split(line).iterator();
-                Map<String, String> row = new HashMap<>(COLUMNS.length);
-
-                while (vals.hasNext() && COLS.hasNext()) {
-                    String col = (String) COLS.next();
-                    String val = vals.next();
-                    row.put(col, val);
-                }
-
-                double lat = Double.parseDouble(row.get("lat"));
-                double lon = Double.parseDouble(row.get("lon"));
-                String rowkey = GeoHash.withCharacterPrecision(lat, lon, 12).toBase32();
-                // ignore duplicates
-                if (!uniqueKeys.contains(rowkey)) {
-                    uniqueKeys.add(rowkey);
-                    Put put = new Put(rowkey.getBytes());
-                    put.setDurability(Durability.SKIP_WAL);
-                    for (Map.Entry<String, String> e : row.entrySet()) {
-                        put.addColumn(FAMILY_A, e.getKey().getBytes(), e.getValue().getBytes());
-                    }
-                    REGION.put(put);
-                    records++;
-                } else {
-                    duplicates++;
-                }
-            }
-        } finally {
-            reader.close();
-        }
-
-        REGION.flush(true);
-
-        long end = System.currentTimeMillis();
-        WIFI_COUNT = records;
-        LOG.info(String.format("Geohashed %s records (%s duplicates) in %sms.", records, duplicates, end - start));
     }
 
     private static void load_rectangle_check_data() throws Exception {
